backend/requirements.txt
Django>=4
djangorestframework
django-cors-headers
Pillow
python-dotenv
psycopg2-binary